Add a makeshift welding blanket
The welding kit requires a welding blanket, but welding blankets are
quite rare and no recipe exists to create one. Doing it properly would
require working with fiberglass or asbestos; instead, add a makeshift
version made from canvas and Nomex, which should be sufficient for
welding in the apocalypse.

Wild guesses made as to material quantities required. Suggestions on
further refinement or added proficiences etc. requested.

resolves #62093